+++ pyriemann/classification.py
@@ -519,7 +519,7 @@ class KNearestNeighbor(MDM):
 
         return self
 
-    def predict(self, covtest):
+    def predict(self, X=None, covtest=None):
         """Get the predictions.
 
         Parameters
@@ -532,7 +532,12 @@ class KNearestNeighbor(MDM):
         pred : ndarray of int, shape (n_matrices,)
             Predictions for each matrix according to the closest centroid.
         """
-        dist = self._predict_distances(covtest)
+        if covtest is not None:
+            print("DeprecationWarning: input covtest has been renamed into X "
+                  "and will be removed in 0.8.0.")
+            X = covtest
+
+        dist = self._predict_distances(X)
         neighbors_classes = self.classmeans_[np.argsort(dist)]
         pred = _mode_2d(neighbors_classes[:, 0:self.n_neighbors], axis=1)
         return pred