+++ doc/whatsnew.rst
@@ -20,6 +20,8 @@ v0.6.dev
 
 - Update to Read the Docs v2. :pr:`260` by :user:`qbarthelemy`
 
+- Deprecate input ``covtest`` for predict of :class:`pyriemann.classification.KNearestNeighbor`, renamed into ``X``. :pr:`259` by :user:`qbarthelemy`
+
 
 v0.5 (Jun 2023)
 ---------------